Review comments new package
