+++ apache-poi-3/src/main/java/com/baeldung/emptyrow/jexcel/JExcelHelper.java
@@ -1,4 +1,4 @@
-package com.baeldung.jexcel;
+package com.baeldung.emptyrow.jexcel;
 
 import jxl.Cell;
