+++ apache-poi-3/src/test/java/com/baeldung/emptyrow/jexcel/JExcelDetectEmptyRowUnitTest.java
@@ -1,6 +1,5 @@
-package com.baeldung.jexcel;
+package com.baeldung.emptyrow.jexcel;
 
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
